requirements.txt
torch>=2.1.0
transformers==4.40.1
pillow==10.3.0
numpy>=1.24.0

beautifulsoup4>=4.12.0
requests>=2.31.0

gradio>=4.18.0